# backend/src/main/resources/API_Calls/dhbw_pawsitters_se2-openapi.yaml
openapi: 3.0.3
info:
  title: Pawsitters API
  version: 1.1.0
  description: OpenAPI-Dokumentation fuer die Pawsitters JSON-API mit einheitlichem Response-Envelope.
  contact:
    name: Pawsitters API Team
servers:
  - url: http://localhost:8080
...
paths:
  /api/auth/register:
    post:
      operationId: authRegister
      tags: [Auth]
      summary: Registrieren und einloggen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registrierung erfolgreich.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/login:
    post:
      operationId: authLogin
      tags: [Auth]
      summary: Einloggen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login erfolgreich.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/users/register:
    post:
      operationId: registerUser
      tags: [Users]
      summary: Benutzer registrieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Benutzer erfolgreich registriert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/users/{id}:
    get:
      operationId: getUserById
      tags: [Users]
      summary: Benutzer abrufen
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateUser
      tags: [Users]
      summary: Benutzer vollstaendig aktualisieren
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Benutzer aktualisiert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      operationId: patchUser
      tags: [Users]
      summary: Benutzer partiell aktualisieren
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatchRequest"
      responses:
        "200":
          description: Benutzer aktualisiert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Benutzer loeschen
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Benutzer geloescht.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteApiResponse"
  /api/users/me:
    get:
      operationId: getCurrentUser
      tags: [Users]
      summary: Aktuellen Benutzer abrufen
      responses:
        "200":
          description: Aktueller Benutzer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/mailExists:
    get:
      operationId: mailExists
      tags: [Users]
      summary: E-Mail pruefen
      parameters:
        - name: mail
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Ergebnis der Pruefung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailExistsApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/users/{id}/roles:
    patch:
      operationId: updateUserRole
      tags: [Users]
      summary: Benutzerrolle aendern
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateRequest"
      responses:
        "200":
          description: Rolle aktualisiert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/users/{id}/profile-image:
    post:
      operationId: uploadUserProfileImage
      tags: [Users]
      summary: Profilbild hochladen
      description: Erwartet ein Multipart-Feld `image`; das Backend akzeptiert aus Kompatibilitaetsgruenden auch `file`.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Profilbild aktualisiert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Ungueltige Anfrage.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
          examples:
            validation:
              $ref: "#/components/examples/ValidationError"
    Unauthorized:
      description: Authentifizierung erforderlich oder fehlgeschlagen.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    Forbidden:
      description: Zugriff verweigert.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    NotFound:
      description: Ressource nicht gefunden.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
  schemas:
    ApiMeta:
      type: object
      required: [timestamp, path]
      properties:
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        total:
          type: integer
          nullable: true
    FieldErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ApiError:
      type: object
      required: [code, details]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldErrorDetail"
...
    ErrorApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required: [error]
          properties:
            success:
              type: boolean
              example: false
            data:
              type: object
              nullable: true
            error:
              $ref: "#/components/schemas/ApiError"
    ApiResponseBase:
      type: object
      required: [success, status, message, meta]
      properties:
        success:
          type: boolean
        status:
          type: integer
        message:
          type: string
        error:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ApiError"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    AuthApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AuthResponse"
...
    UserApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/UserResponse"
...
    MailExistsApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MailExistsResponse"
    DeleteApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/DeleteResponse"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, phone, birthDate, emergencyContact, profilePicture, bio, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        emergencyContact:
          type: string
        profilePicture:
          type: string
        bio:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    UserUpdateRequest:
      type: object
      required: [firstName, lastName, phone, birthDate, emergencyContact, profilePicture, bio]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        emergencyContact:
          type: string
        profilePicture:
          type: string
        bio:
          type: string
    UserPatchRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        emergencyContact:
          type: string
        profilePicture:
          type: string
        bio:
          type: string
    RoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    AuthResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string
...
    MailExistsResponse:
      type: object
      required: [exists]
      properties:
        exists:
          type: boolean
    DeleteResponse:
      type: object
      required: [deleted, id, message]
      properties:
        deleted:
          type: boolean
        id:
          type: integer
          format: int64
        message:
          type: string
    UserResponse:
      type: object
      description: Oeffentliche User-Daten. passwordHash wird nie ausgeliefert.
      required: [id, email, firstName, lastName, phone, birthDate, emergencyContact, profilePicture, bio, rating, numberOfRatings, role, pets]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        emergencyContact:
          type: string
        profilePicture:
          type: string
        bio:
          type: string
        rating:
          type: number
          format: float
        numberOfRatings:
          type: integer
          format: int32
        role:
          $ref: "#/components/schemas/UserRole"
        pets:
          type: array
          items:
            $ref: "#/components/schemas/PetResponse"
    PetResponse:
      type: object
      required: [id, name, species, breed, age, specialNeeds, defaultImagePath]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        species:
          $ref: "#/components/schemas/PetChoice"
        breed:
          type: string
        age:
          type: integer
          format: int32
        specialNeeds:
          type: string
        imagePath:
          type: string
          nullable: true
        defaultImagePath:
          type: string
    UserRole:
      type: string
      enum: [PET_OWNER, HOST, ADMIN]
    PetChoice:
      type: string
      enum: [DOG, CAT, RABBIT, HAMSTER, GUINEA_PIG, PARROT, BUDGIE, CANARY, TURTLE, SNAKE, LIZARD, FERRET, RAT, MOUSE, FISH, HORSE, DONKEY, GOAT, CHICKEN, BIRD]
  examples:
...
    ValidationError:
      summary: Validierungsfehler
      value:
        success: false
        status: 400
        message: Validation failed.
        data: null
        error:
          code: VALIDATION_FAILED
          details:
            - field: email
              message: must be a well-formed email address
        meta:
          timestamp: "2026-05-06T19:30:00Z"
          path: /api/auth/register